+++ Readme.md
@@ -1,7 +1,7 @@
 plusieurs arg sous cette forme 
 arg1 "I" arg2 "I" arg3
 
-fait depuis samedi :
+fait samedi :
 
 re creation des arguments en tableau de liste
 
@@ -12,6 +12,8 @@ recoder les childs pipex
 
 struct des pipes et modification de ces derniers entre chaque appel de child
 
+fait lundi :
+
 creer un systeme agile pour les redirections
 	-idee : au lieu de faire 3 childs distinct redirige en fonction des sorties 
 	- 4 possibilite : fd_in et fd_out
@@ -19,8 +21,16 @@ creer un systeme agile pour les redirections
 					  pipe_in et pipe_out
 					  fd_in et pipe_out
 
+prise en compte des builtins
+
+update commentaire
+
+rangement des fonctions dans un systeme de fichier logique et bien organise
+
 a faire :
 
-free + close tout les pipes et fd
-
+free + close tout les pipes et fd dans le cas ou tout est ok
+free + close tout les pipes et fd dans les cas d'erreur
+retour du $?
+norme
 signaux ?
