+++ header/minishell.h
@@ -6,7 +6,7 @@
 /*   By: lmarchai <lmarchai@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/30 14:39:47 by lmarchai          #+#    #+#             */
-/*   Updated: 2023/09/18 08:33:52 by lmarchai         ###   ########.fr       */
+/*   Updated: 2023/09/18 10:04:57 by lmarchai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -66,8 +66,6 @@ typedef struct s_cmd
     int                    fd_out;
 }      t_cmd;
 
-
-void	child_process(t_cmd *cmd, t_pipe *pipes, char **envp);
 char	*get_args(char **argv);
 char	*get_cmd(char **paths, char *cmd);
 char	*nopath(char *cmd);
@@ -75,4 +73,25 @@ char	*find_path(char **envp);
 char	**gen_first_cmd(char **argv);
 char	**gen_sec_cmd(char **argv);
 char	**gen_third_cmd(char **argv, int argc);
+
+int		strlen_list(t_cmd *cmd[4]);
+
 void	cross_array_list(t_cmd *cmd[4], char **envp);
+t_pipe	*gen_child(t_cmd *cmd, t_pipe *pipes, char **envp, int i);
+t_pipe	*new_pipes(t_pipe *pipes, int i);
+void	wait_childs(t_cmd *cmd[4]);
+void	child_process(t_cmd *cmd, t_pipe *pipes, char **envp);
+
+t_pipe	*handle_redirection(t_cmd *cmd, t_pipe *pipes);
+void	handle_builtins(t_cmd *cmd, t_pipe *pipes, char **envp);
+
+void	free_list_args(t_cmd *cmd[4], t_pipe *pipes, int len_list);
+
+void	close_pipes(t_pipe *pipes);
+
+int		strlen_list(t_cmd *cmd[4]);
+
+void	error_pipe(void);
+void	error_dup2(void);
+void	error_malloc(void);
+void	error_fork(void);
