+++ srcs/close.c
@@ -1,0 +1,21 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   close.c                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: lmarchai <lmarchai@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/09/18 09:53:08 by lmarchai          #+#    #+#             */
+/*   Updated: 2023/09/18 09:53:33 by lmarchai         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../header/minishell.h"
+
+void	close_pipes(t_pipe *pipes)
+{
+	close(pipes->tube[0][0]);
+	close(pipes->tube[0][1]);
+	close(pipes->tube[1][0]);
+	close(pipes->tube[1][1]);
+}
